atacseq schema & QualityMetricView updates
